move utils in new file
