+++ src/utils.ts
@@ -1,0 +1,43 @@
+import { Operation } from '@apollo/client/core'
+import { OperationDefinitionNode, StringValueNode } from 'graphql'
+
+export const getDirectiveArgumentValueFromOperation = (
+  operation: Operation,
+  directiveName: string,
+  argumentName: string
+) =>
+  ((operation.query.definitions.find(
+    (definition) => definition.kind === 'OperationDefinition'
+  ) as OperationDefinitionNode)?.directives
+    ?.find((directive) => directive.name?.value === directiveName)
+    ?.arguments?.find((argument) => argument.name?.value === argumentName)
+    ?.value as StringValueNode)?.value
+
+export const addApiNameToTypeName = (data: any, apiName: string): any => {
+  if (data == null || typeof data !== 'object') {
+    return data
+  }
+  if (Array.isArray(data)) {
+    return data.map((item) => addApiNameToTypeName(item, apiName))
+  }
+
+  const newData = Object.entries(data).reduce(
+    (ctx, [itemKey, item]) => ({
+      ...ctx,
+      [itemKey]: addApiNameToTypeName(item as any, apiName),
+    }),
+    {}
+  )
+
+  if (data.__typename) {
+    return {
+      ...newData,
+      __typename: `${apiName}:${data.__typename}`,
+    }
+  }
+
+  return newData
+}
+
+export const isFunction = (fn: any) =>
+  fn && {}.toString.call(fn) === '[object Function]'
